pyproject.toml
[tool.poetry]
name = "mystories"
version = "0.1.0"
description = ""
authors = ["npujol <naivy.luna@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django-rest-framework = "^0.1.0"
django-cors-middleware = "^1.5.0"
django-extensions = "^2.2.9"
pyjwt = "^1.7.1"
six = "^1.15.0"
drf-yasg = "^1.17.1"
gtts = "^2.1.1"
celery = "^4.4.5"
django-celery-results = "^1.2.1"
django-redis = "^4.12.1"
apprise = "^0.8.6"
psycopg2 = "^2.8.5"
flower = "^0.9.4"
djangorestframework_simplejwt = "^4.4.0"
pillow = "^7.2.0"
drf-nested-routers = "^0.91"
django-filter = "^2.3.0"
markdown = "^3.2.2"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
